pyproject.toml
[project]
name = "canva-workshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Luca Picci",email = "lpicci96@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python =">=3.11,<4.0"
dependencies = [
    "bblocks-data-importers @ git+https://github.com/ONEcampaign/bblocks_data_importers.git",
    "pandas (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
